Split MiniParseEventSource into multiple event sources
